Adding Cardiff

+++ catscripts/Technology/Computers/Software/DVD_authoring/DVDauthoring.sh
@@ -1,22 +1,23 @@
 #!/bin/bash
 
-egrep -i 'ISO (image|file)|(CD|DVD|Blu(-|| )Ray) authoring' newpages.txt >> Authoring.txt
-egrep -i 'Nero Burning ROM' newpages.txt >> Authoring.txt #CD/DVD writing software
-egrep -i 'Nero Linux' newpages.txt >> Authoring.txt 
-egrep -i 'burn a (CD|DVD|Blu(|-| )Ray)' newpages.txt >> Authoring.txt
-egrep -i 'K3B|Burn Baby Burn' newpages.txt >> Authoring.txt #Linux CD/DVD burning software
-egrep -i 'ISO(| )Master' newpages.txt >> Authoring.txt #ISO image creator
-egrep -i 'ImgBurn' newpages.txt >> Authoring.txt
-egrep -i 'CDBurnerXP' newpages.txt >> Authoring.txt
-egrep -i 'InfraRecorder' newpages.txt >> Authoring.txt
+KEYWORDS_DVDAUTHORING="ISO (image|file)|(CD|DVD|Blu(-|| )Ray)(| )authoring|burn a (CD|DVD|Blu(|-| )Ray)|Burn(| )(CD|DVD)(|s)"
+KEYWORDS_DVDAUTHORING_OTHER="Nero(| )(Linux|Burning(| )ROM)|K3B|Burn(| )Baby(| )Burn|ISO(| )Master|ImgBurn|DVD(| )Decrypter|CDBurner(| )XP|Infra(| )Recorder" #Names of apps that don't currently have their own category
 
-AUTHORING=`stat --print=%s Authoring.txt`
-
-if [ $AUTHORING -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Authoring.txt"
-  export CATNAME="CD and DVD authoring"
-  $CATEGORIZE
-fi
 
-rm Authoring.txt
+  egrep -i "$KEYWORDS_DVDAUTHORING" newpages.txt >> Authoring.txt
+  egrep -i "$KEYWORDS_DVDAUTHORING_OTHER" newpages.txt >> Authoring.txt
+
+  AUTHORING=`stat --print=%s Authoring.txt`
+
+  if [ $AUTHORING -ne 0 ];
+  then
+    export CATFILE="Authoring.txt"
+    export CATNAME="CD and DVD authoring"
+    $CATEGORIZE
+  fi
+
+  rm Authoring.txt
+
+fi
